#121: Fixed issue with list/map size() so it now reverts to returning int values rather than longs

+++ src/main/java/io/jactl/runtime/BuiltinFunctions.java
@@ -920,9 +920,9 @@ public class BuiltinFunctions {
 
   // = size
 
-  public static long mapSize(Map map) { return map.size(); }
+  public static int mapSize(Map map) { return map.size(); }
 
-  public static long listSize(List list) {
+  public static int listSize(List list) {
     return list.size();
   }
